pyproject.toml
[tool.poetry]
name = "azdummy"
version = "0.0.0"
description = "A Python Typer-based CLI tool to generate fake data for Azure AD."
authors = ["Leron Gray <daddycocoaman@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^9.3.0"
pydantic = "^1.7.3"
python-dotenv = "^0.15.0"
mimesis = "^4.1.2"
click-help-colors = "^0.9"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
black = "^20.8b1"
mkdocs = "^1.1.2"
mkdocs-click = "^0.2.0"
mkdocs-material = "^6.1.7"
typer-cli = "^0.0.11"

[tool.poetry.scripts]
azdummy = "azdummy.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
